Add files via upload
This application focus on table/Order booking and management of details, so customers can concentrate on what is important for them. The users and stakeholders for this application will be customers as it allows them to reserve a table within seconds by choosing their required venue and number of persons accompanying them within their budget. Customers can only reserve their table by selecting food and venue. The details of their booking can be viewed in booking history.